+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - (Feature) (Platform) Shutdown migration to CE
 - (Feature) (Scheduler) Shutdown Integration
 - (Feature) CertManager Integration
+- (Feature) (Networking) Gateway Options sync
 
 ## [1.2.43](https://github.com/arangodb/kube-arangodb/tree/1.2.43) (2024-10-14)
 - (Feature) ArangoRoute CRD